Fix issues #555, #556, #557, #558: Changesets, test isolation, mock separation, bundle analysis
- #558: Add changesets workflow documentation and CI enforcement
- #557: Verify useBalance test cache reset (already properly implemented)
- #555: Separate DashboardStats types from mock data, create real API client
- #556: Add bundle analyzer to build pipeline with CI artifact upload
