+++ .changeset/initial-v0.1.0.md
@@ -1,0 +1,16 @@
+---
+"clipsproject": minor
+---
+
+Initial v0.1.0 release with core features:
+- Next.js 16 application with TypeScript
+- Zustand state management
+- Stellar wallet integration
+- Multi-wallet support
+- Social recovery system
+- Dashboard with earnings tracking
+- User authentication with NextAuth
+- Storybook component documentation
+- Jest unit tests and Playwright E2E tests
+- Sentry error monitoring
+- Security sanitization with DOMPurify
