+++ app/hooks/useBalance.test.ts
@@ -359,11 +359,19 @@ describe("useBalance", () => {
     };
 
     (global.fetch as jest.Mock)
-      .mockResolvedValue({
+      .mockResolvedValueOnce({
         ok: true,
         json: async () => mockAccountData,
       })
-      .mockResolvedValue({
+      .mockResolvedValueOnce({
+        ok: true,
+        json: async () => mockPriceData,
+      })
+      .mockResolvedValueOnce({
+        ok: true,
+        json: async () => mockAccountData,
+      })
+      .mockResolvedValueOnce({
         ok: true,
         json: async () => mockPriceData,
       });
@@ -381,11 +389,6 @@ describe("useBalance", () => {
     });
 
     const firstFetchTime = result.current.lastFetchTime;
-
-    // Wait a bit
-    await act(async () => {
-      await new Promise((resolve) => setTimeout(resolve, 100));
-    });
 
     // Manual refresh
     await act(async () => {
@@ -407,11 +410,19 @@ describe("useBalance", () => {
     };
 
     (global.fetch as jest.Mock)
-      .mockResolvedValue({
+      .mockResolvedValueOnce({
         ok: true,
         json: async () => mockAccountData,
       })
-      .mockResolvedValue({
+      .mockResolvedValueOnce({
+        ok: true,
+        json: async () => mockPriceData,
+      })
+      .mockResolvedValueOnce({
+        ok: true,
+        json: async () => mockAccountData,
+      })
+      .mockResolvedValueOnce({
         ok: true,
         json: async () => mockPriceData,
       });
@@ -448,11 +459,17 @@ describe("useBalance", () => {
   });
 
   it("should clear error", async () => {
-    (global.fetch as jest.Mock).mockResolvedValueOnce({
-      ok: false,
-      status: 404,
-      statusText: "Not Found",
-    });
+    (global.fetch as jest.Mock)
+      .mockResolvedValueOnce({
+        ok: false,
+        status: 404,
+        statusText: "Not Found",
+      })
+      .mockResolvedValueOnce({
+        ok: false,
+        status: 404,
+        statusText: "Not Found",
+      });
 
     const { result } = renderHook(() =>
       useBalance({
@@ -493,14 +510,14 @@ describe("useBalance", () => {
       });
 
     const { result, rerender } = renderHook(
-      ({ publicKey }) =>
+      ({ publicKey }: { publicKey: string | null }) =>
         useBalance({
           publicKey,
           network: "TESTNET",
           autoRefresh: false,
         }),
       {
-        initialProps: { publicKey: "GTEST123" },
+        initialProps: { publicKey: "GTEST123" as string | null },
       }
     );
