+++ next.config.ts
@@ -1,5 +1,6 @@
 import type { NextConfig } from "next";
 import { validateRequiredEnv } from "./app/lib/validateEnv";
+import withBundleAnalyzer from "@next/bundle-analyzer";
 
 validateRequiredEnv();
 
@@ -20,6 +21,28 @@ const nextConfig: NextConfig = {
       },
     ],
   },
+  webpack: (config: any, { isServer }: { isServer: boolean }) => {
+    if (!isServer) {
+      config.optimization = {
+        ...config.optimization,
+        splitChunks: {
+          chunks: 'all',
+          cacheGroups: {
+            default: false,
+            vendors: false,
+            commons: {
+              name: 'commons',
+              chunks: 'all',
+              minChunks: 2,
+            },
+          },
+        },
+      };
+    }
+    return config;
+  },
 };
 
-export default nextConfig;
+export default withBundleAnalyzer({
+  enabled: process.env.ANALYZE === 'true',
+})(nextConfig);
