+++ package.json
@@ -12,7 +12,8 @@
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "prepare": "husky install",
-    "changeset": "changeset"
+    "changeset": "changeset",
+    "analyze": "ANALYZE=true next build"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.637.0",
@@ -35,6 +36,7 @@
   "devDependencies": {
     "@changesets/cli": "^2.27.0",
     "@chromatic-com/storybook": "^5.2.1",
+    "@next/bundle-analyzer": "^16.2.9",
     "@playwright/test": "^1.59.1",
     "@storybook/addon-a11y": "^10.4.1",
     "@storybook/addon-docs": "^10.4.1",
